+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v0.115.2
+- Removed the untested version warning for AI 2023.
+- Addressed a run-time error thrown when trying to modify a settings text block that is hidden or locked.
+
 ### v0.115.1
 - Added "playsinline" attribute to video underlays, to fix a mobile Safari issue.